move transit model to its own file
